Build properties and methods with name so they are merged
Magic properties are less relevant when a native property is defined.
We will ignore the tags when defined if a native property is present.
