+++ src/phpDocumentor/Descriptor/ClassDescriptor.php
@@ -13,16 +13,9 @@ declare(strict_types=1);
 
 namespace phpDocumentor\Descriptor;
 
-use InvalidArgumentException;
 use phpDocumentor\Descriptor\Interfaces\AttributedInterface;
 use phpDocumentor\Descriptor\Interfaces\ClassInterface;
-use phpDocumentor\Descriptor\Interfaces\MethodInterface;
-use phpDocumentor\Descriptor\Interfaces\PropertyInterface;
-use phpDocumentor\Descriptor\Validation\Error;
 use phpDocumentor\Reflection\Fqsen;
-
-use function ltrim;
-use function sprintf;
 
 /**
  * Descriptor representing a Class.
@@ -54,84 +47,6 @@ class ClassDescriptor extends DescriptorAbstract implements Interfaces\ClassInte
     public function isReadOnly(): bool
     {
         return $this->readOnly;
-    }
-
-    /** @return Collection<MethodInterface> */
-    public function getMagicMethods(): Collection
-    {
-        $methodTags = $this->getTags()->fetch('method', new Collection())->filter(Tag\MethodDescriptor::class);
-
-        $methods = Collection::fromInterfaceString(MethodInterface::class);
-
-        foreach ($methodTags as $methodTag) {
-            $method = new MethodDescriptor();
-            $method->setName($methodTag->getMethodName());
-            $method->setDescription($methodTag->getDescription());
-            $method->setStatic($methodTag->isStatic());
-            $method->setParent($this);
-            $method->setReturnType($methodTag->getResponse()->getType());
-            $method->setHasReturnByReference($methodTag->getHasReturnByReference());
-
-            $returnTags = $method->getTags()->fetch('return', new Collection());
-            $returnTags->add($methodTag->getResponse());
-
-            foreach ($methodTag->getArguments() as $name => $argument) {
-                $method->addArgument($name, $argument);
-            }
-
-            $methods->add($method);
-        }
-
-        $parent = $this->getParent();
-        if ($parent instanceof static) {
-            $methods = $methods->merge($parent->getMagicMethods());
-        }
-
-        return $methods;
-    }
-
-    /** @return Collection<PropertyInterface> */
-    public function getMagicProperties(): Collection
-    {
-        $tags = $this->getTags();
-        /** @var Collection<Tag\PropertyDescriptor> $propertyTags */
-        $propertyTags = $tags->fetch('property', new Collection())->filter(Tag\PropertyDescriptor::class)
-            ->merge($tags->fetch('property-read', new Collection())->filter(Tag\PropertyDescriptor::class))
-            ->merge($tags->fetch('property-write', new Collection())->filter(Tag\PropertyDescriptor::class));
-
-        $properties = Collection::fromInterfaceString(PropertyInterface::class);
-
-        /** @var Tag\PropertyDescriptor $propertyTag */
-        foreach ($propertyTags as $propertyTag) {
-            $property = new PropertyDescriptor();
-            $property->setName(ltrim($propertyTag->getVariableName(), '$'));
-            $property->setDescription($propertyTag->getDescription());
-            $property->setType($propertyTag->getType());
-            $property->setWriteOnly($propertyTag->getName() === 'property-write');
-            $property->setReadOnly($propertyTag->getName() === 'property-read');
-            try {
-                $property->setParent($this);
-                $properties->add($property);
-            } catch (InvalidArgumentException $e) {
-                $this->errors->add(
-                    new Error(
-                        'ERROR',
-                        sprintf(
-                            'Property name is invalid %s',
-                            $e->getMessage(),
-                        ),
-                        null,
-                    ),
-                );
-            }
-        }
-
-        $parent = $this->getParent();
-        if ($parent instanceof self) {
-            $properties = $properties->merge($parent->getMagicProperties());
-        }
-
-        return $properties;
     }
 
     /** @return ClassInterface|Fqsen|string|null */
